backend/requirements.txt
fastapi
uvicorn
pydantic
joblib
pandas
numpy
catboost
scikit-learn
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
python-dotenv
google-generativeai
requests
